refactor：adjust dir; optimise route change
