+++ src/logseq/logseqPluginInit.tsx
@@ -11,6 +11,7 @@ export const refreshLogseqInit = () => {
 // 打开文件管理器的函数
 export const openFileManager = () => {
     logger.debug(`openFileManager`)
+
     logseq.showMainUI();
 };
 
@@ -40,17 +41,16 @@ export const setupStyles = () => {
       @import url('https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0');
     `);
     logseq.setMainUIInlineStyle({
-        zIndex: 1000,
-    });
+        zIndex: 1,
+    }); 
 };
 
 // 设置文件管理器导航
 export const setupFileManagerNav = async () => {
     const { preferredLanguage } = await logseq.App.getUserConfigs()
-    const fileManagerDiv = document.createElement('div');
-    fileManagerDiv.id = 'fileManager-nav-id'
+    const fileManagerDiv = document.createElement('div'); 
     fileManagerDiv.innerHTML = `
-      <a class='item group flex items-center text-sm font-medium rounded-md fileManager-nav'>
+      <a href="#/page/files-manager" class='item group flex items-center text-sm font-medium rounded-md'>
         <span class='ui__icon ti ls-icon-calendar'>
           <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 256 256'><rect width='20' height='20' fill='none'/><rect x='32' y='48' width='192' height='160' rx='8' fill='none' stroke='currentColor' stroke-linecap='round' stroke-linejoin='round' stroke-width='16'/><circle cx='68' cy='84' r='12'/><circle cx='108' cy='84' r='12'/></svg>      </span>
         <span class='flex-1'>${getI18nConstant(preferredLanguage, i18n_FILE_MANAGER_LABEL)}</span>
@@ -66,17 +66,7 @@ export const setupFileManagerNav = async () => {
         navHeader.insertBefore(fileManagerDiv, navHeader.lastChild);
     }
 };
-
-// 处理点击事件，点击外部时关闭主界面
-export const handleClickOutside = () => {
-    document.body.addEventListener('click', (e) => {
-        if ((e.target as HTMLElement).classList.length === 0) {
-            logseq.hideMainUI({ restoreEditingCursor: true });
-        }
-    });
-
-    document.getElementById('app')?.addEventListener('click', e => e.stopPropagation());
-};
+ 
 
 export function doInitNotifyUser(showMsg: () => void) {
     logger.debug('logseq.updateSettings', logseq.settings, 'logseq.baseInfo.version', logseq.baseInfo.version)
