+++ src/ui_hooks/useCalculateHeight.tsx
